# feat: Support `PiecewiseMergeJoin` to speed up single range predicate joins