--- a/datafusion/physical-plan/src/joins/mod.rs
+++ b/datafusion/physical-plan/src/joins/mod.rs
@@ -23,12 +23,14 @@ use datafusion_physical_expr::PhysicalExprRef;
 pub use hash_join::HashJoinExec;
 pub use nested_loop_join::NestedLoopJoinExec;
 use parking_lot::Mutex;
+pub use piecewise_merge_join::PiecewiseMergeJoinExec;
 // Note: SortMergeJoin is not used in plans yet
 pub use sort_merge_join::SortMergeJoinExec;
 pub use symmetric_hash_join::SymmetricHashJoinExec;
 mod cross_join;
 mod hash_join;
 mod nested_loop_join;
+mod piecewise_merge_join;
 mod sort_merge_join;
 mod stream_join_utils;
 mod symmetric_hash_join;
@@ -67,5 +69,5 @@ pub enum StreamJoinPartitionMode {
     SinglePartition,
 }
 
-/// Shared bitmap for visited left-side indices
+/// Shared bitmap for visited indices
 type SharedBitmapBuilder = Mutex<BooleanBufferBuilder>;