@jonathanc-n 2025-07-05
cc @alamb @ozankabak Seems like you guys were part of the discussion for range joins, this is a nice start to it? @Dandandan @comphead might be interested?