@comphead 2025-07-07
Thanks @jonathanc-n let me first get familiar with this kind of join