@alamb 2025-08-07
Thanks @jonathanc-n -- unfortunately I am not likely to have time to review this as my focus hasn't been on the join implementation

@comphead do you know anyone who is more focused joins these days who might be able to help review this?