@comphead 2025-08-07
Hey, sorry I missed this, this join is quite interesting concept, I'm planning to finish review #16996 this week, and switch to this PR next.