@2010YOUY01 2025-08-08
Thanks for this great work, as always! I got some high-level questions.

Is it possible that IE Join could be faster for the same workload where PMG performs well? Should we implement IE Join directly? (I’m willing to help review or implement it at this point)

There was an old attempt (https://github.com/apache/datafusion/pull/12754), but unfortunately it didn’t get reviewed at the time.