@jonathanc-n 2025-08-11
> Support additional join predicates https://github.com/apache/datafusion/pull/16660#discussion_r2265310584
I dont think this should be done. I looked into it, and the overhead + complexity brought by checking selectivity just for a workload which is very unlikely (equijoin filters less than the single filter). DuckDB also doesnt support this if that is an indicator at all.