@jonathanc-n 2025-08-12
@2010YOUY01 Wouldn't you need to use a cost model to estimate which one to use though when both are viable? For example, the Hash Join (do the equi-condition then the residual filter) vs. PWMJ (do the filter, then the equi condition residual). You could esimate the selectivity for the equi predicate vs. residual predicate, and also factor in whether the key is sorted. 


Sorry if i misinterpreted this