@2010YOUY01 2025-08-13
> Yes I think so too, I don't know if it will be worth the complexity though since this is a very niche workload (single range filter + higher selectivity for filter)
> 
> I think the plan for now is that I can implement support for AND expressions in this pull request. But when we include the planner changes in a follow up PR we can discuss this there.

I think in-the-wild join workloads most commonly involve lengthy ANDed expressions, so it’s indeed challenging to make a smart planner.
It’s a good idea to have this PR focus on execution, and handle the planning in another PR.